# Restore #1604: Use getAncestors in getDescendant