PR #1589 overwrites #1604 PR.  I think it is because git text comparison fails to warn the conflict between PR 1589 and other PRs.  

This PR is to restore PR #1604 , which is to use getAncestors in getDescendant to speed up the getDescendant when getAncestors is cached.