@ianstormtaylor 2018-03-21
Why? It doesn't appear to make things faster?