@zhujinxuan 2018-03-22
I think there are two reasons.  
1. using getAncestors get code shorter
2. it is normal in plugins to use getClosestXXX or other ancestors related methods after getDescendants.  For example, whether a text is in void inline, void image or table.

And a potential reason I am not sure about:
I observe a strange benchmark speed up:  getLeaves is speed up when both this PR and the post interface of memorize is added. 

Let us postpone it.  I will add some benchmark files about delete/insert with nested structures to see whether this PR is good.