@ianstormtaylor 2018-03-22
Awesome, that sounds good to me. Thanks!

If we do decide we want to go with this technique, then I think it would be worth adding a `getDescendants(key)` or something similar that returns ancestors **and** the child itself. Then `getAncestors` and `getDescendant` and anything else can all call into a single method. Only reason I say that is because I think it's more clear than using `getAncestors().last().getChild(key)` which is hard to follow.