@zhujinxuan 2018-04-08
Close this PR; Because https://github.com/ianstormtaylor/slate/pull/1763 is a much better idea ~